ACTUALLLLLY...now i`m finished
decided to fix 1-2 elements regarding images and window sizes. The only regret i have in this project is that i`m not able to link another one of my projects home page onto it. I`m under the assumption that i would need to include the website`s html into the static folder, but for now, the project fully works and is good to upload.